Fix mocking of Online::Queue
As it turns out, we weren't correctly mocking Online::Queue.  This error
had slipped through because we were always testing with a network, and
the calls were simply being passed through to S3.  This patch configures
webmock to block all network access when we're supposed to be mocking
our interfaces, and a makes a moderately long list of changes to get
everything working correctly.

We also change the name of test buckets used when testing the queue to
include the user's login, just as with all other kinds of test buckets.
This is more consistent, and it will prevent tests from stomping on each
other when run simultaneously by multiple users.